Reorganize authentication-related endpoints.
